* development:
  fix: db error when starting state sync fixes #225 (#227)
  feat: add template name to list templates + minor cleanups (#226)
  fix: remove handle new valid transaction check (#224)
  fix: vns should not store transactions for shards outside committee range (see issue #193) (#206)
  feat!: add is_mut field to function def (see issue #213) (#214)
  fix: fix VN merkle root mismatch (#220)
  feat: determine leader from payload (#221)

+++ dan_layer/storage_sqlite/src/global/models/epoch.rs
@@ -26,7 +26,7 @@ use crate::global::schema::*;
 
 #[derive(Queryable)]
 pub struct Epoch {
-    pub epoch: i32,
+    pub epoch: i64,
     pub validator_node_mr: Vec<u8>,
 }
 
@@ -42,14 +42,14 @@ impl From<Epoch> for DbEpoch {
 #[derive(Insertable)]
 #[table_name = "epochs"]
 pub struct NewEpoch {
-    pub epoch: i32,
+    pub epoch: i64,
     pub validator_node_mr: Vec<u8>,
 }
 
 impl From<DbEpoch> for NewEpoch {
     fn from(e: DbEpoch) -> Self {
         Self {
-            epoch: e.epoch as i32,
+            epoch: e.epoch as i64,
             validator_node_mr: e.validator_node_mr,
         }
     }
